+++ Spooksoft.Xml.Serialization.Test/Models/Simple/DerivedPropertyType2.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Spooksoft.Xml.Serialization.Test.Models.Simple
+{
+    public class DerivedPropertyType2 : BasePropertyType
+    {
+        public float FloatProperty { get; set; }
+    }
+}