+++ Changelog.md
@@ -7,3 +7,8 @@
 * [new] Added support for single-dimensional arrays
 * [new] Added support for maps (e.g. Dictionary<,>)
 * [new] Added support for binary serialization (only for byte[] properties)
+
+# 1.0.2
+
+* [new] Added support for varying-type properties
+* Refactored implementation of varying-type item serialization (XmlArrayItem, XmlMapKeyItem, XmlMapValueItem, XmlVariant) to reduce duplicated code and simplify implementation